gh-96859: [argparse] Clone each destination list only once
Currently the append, append_const, and extend actions make a copy of the destination
list every time the action is called. This is noteceable with very long option lists.

This commit adds Action.clone_dest() method that is called the first time an action
runs for each destination. This allows to copy lists only once, significantky speeding
up appends.